+++ engines/lastexpress/game/action.cpp
@@ -394,7 +394,7 @@ Action::Action(LastExpressEngine *engine) : _engine(engine) {
 }
 
 Action::~Action() {
-	for (int i = 0; i < (int)_actions.size(); i++)
+	for (uint i = 0; i < _actions.size(); i++)
 		SAFE_DELETE(_actions[i]);
 
 	_actions.clear();
